refactor(all): new squema for the api
Add token and basic authorization by default

+++ reset_db.sh
@@ -1,9 +1,13 @@
-#/bin/bash
+#!/bin/bash
 # Borrar migraciones de cada carpeta
 
 rm db.sqlite3
-# Borrar migraciones de cada carpeta
-for app in login suscripciones control_vehicular operaciones; do
-    find ./$app/migrations -type f ! -name '__init__.py' -delete
+# Obtener una lista de todas las carpetas en el directorio actual
+for app in $(ls -d */); do
+    # Excluir las carpetas administracion_vehicular y venv
+    if [ "$app" != "administracion_vehicular/" ] && [ "$app" != "venv/" ]; then
+        # Borrar migraciones de la carpeta actual
+        find ./$app/migrations -type f ! -name '__init__.py' -delete
+    fi
 done
 ./migrate.sh
